Some modifications to the Spanish Webpage translation.
These changes are to make the current Spanish Web pages
more neutral. Currently it has lots of terminology that is used in
Spain and not in South America like:

Ficheros we use Archivos
Ordenador we use Computador
Podeis we use Puedes

Plus some accents missing here and there and typos.

Signed-off-by: Pablo Wolter <pwolter@gmail.com>
Signed-off-by: Dirk Hohndel <dirk@hohndel.org>